# mathapp.apiDocs.yaml
openapi: '3.0.0'
info:
  version: "0.1.0"
  title: 'MathApp.API'
  description: 'This is the backend API for the ASU STEM Academy Math APP Project'
servers:
  - url: /api/v1
paths:
  /students:
    get:
      description: Returns a list of students
      operationId: students.get_all
      responses:
        '200':
          description: Student List
          content:
            application/json:
              schema:
                type: object
                properties:
                  students:
                    type: array
                    items:
                      $ref: '#/components/schemas/student'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
    post:
      description: Creates a new student record
      operationId: students.add_student
      requestBody:
        description: Student to add
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/new_student'
      responses:
        '201':
          description: Created
          headers:
            Location:
              schema:
                type: string
              description: Record Locator
        '400':
          $ref: '#/components/responses/400Error'
  /students/{student_id}:
    get:
      description: Get student by student ID number
      operationId: students.get_student
      parameters:
        - name: student_id
          description: Student ID to lookkup
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Student Record
          content:
            application/json:
              schema:
                properties:
                  student:
                    type: object
                    $ref: '#/components/schemas/student'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
...
  /results/student/{student_id}:
    get:
      description: Get results for all tests by the specified student
      operationId: results.get_by_student
      parameters:
        - name: student_id
          description: Student ID to lookup
          in: path
          schema:
            type: integer
            minimum: 1
          required: true
      responses:
        '200':
          description: Aggregate Results
          content:
            application/json:
              schema:
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/result'
        '400':
          $ref: '#/components/responses/400Error'
        '404':
          $ref: '#/components/responses/404Error'
...
components:
  schemas:
    student:
      type: object
      properties:
        id:
          type: integer
          description: Student ID
        student_firstname:
          type: string
          description: Student's First Name
        student_lastname:
          type: string
          description: Student's Last Name
        teacher_id:
          type: string
          description: Student's Teacher
        stars:
          type: integer
          description: Stars student has earned
        level:
          type: integer
          description: Level of student
...
    result:
      type: object
      properties:
        result_id:
          type: integer
          description: Tracking number for test attempt submission
        time_taken:
          type: string
          description: Time taken for the individual test attempt ("HH:MM:SS")
        testId:
          type: integer
          description: Test ID of test this submission is for
        answers:
          type: array
          items:
            properties:
              student_answer:
                type: integer
                description: Answer student submitted 
              correctly_answered:
                type: boolean
                description: Whether the student's answer was correct
              operation:
                type: string
                description: +, -, *, /
              operand_1:
                type: integer
                description: Operand preceding the operation
              operand_2:
                type: integer
                description: Operand subequent to operation
...
    error:
      type: object
      properties:
        message:
          type: string
          description: Error Message from server
    new_student:
      type: object
      properties:
        student_firstname:
          type: string
          description: Student's First Name
        student_lastname:
          type: string
          description: Student's Last Name
        teacher_id:
          type: integer
          description: ID for Student's Teacher
          default: 0
      required:
        - student_firstname
        - student_lastname
        - teacher_id
...
  responses:
    400Error:
      description: Invalid Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    404Error:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'